Extract the Session module (mint, outcomes, SSO-only) from the flat kernel
Continues the kernel dissolution (#790): the session-minting and login-outcome
types leave the flat Api namespace into Jellyfin.Plugin.SSO_Auth.Api.Session:

- SessionMinter, SessionParameters, LoginOutcome, PublicReason, AuthResponse,
  LoginStatusMapper, SsoOnlyLoginService, SsoOnlyReconciliationService,
  SsoAuthenticationProviders -> Api/Session/ (namespace Api.Session), import
  added at each call site.
- DAG edges: Session -> {Authz, Avatar, Linking} (grants, avatar set at mint,
  link reconciliation). Reverse edge declared: Saml -> Session
  (SamlAssertionValidator returns LoginOutcome/PublicReason). The old
  SessionMinter <-> Avatar coupling turned out to be comment-only, so the edge
  is one-way and clean.
- Test mirror: the seven session-behaviour test files move to Tests/Session/.
- The SessionMinter_RechecksRevocationImmediatelyBeforeTheMint conformance rule
  read SessionMinter.cs by a hardcoded flat-Api path; updated to the new
  location (RCA: the migration playbook's landmine check covered opengrep rules
  but not hardcoded file paths inside the conformance tests themselves — the
  check now covers both; no other hardcoded Api file path remains).

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1585 both TFMs incl. the
module DAG and mirror rules), and the ABI-floor build.

Refs #790
