+++ SSO-Auth/Api/Session/PublicReason.cs
@@ -1,4 +1,4 @@
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Session;
 
 /// <summary>
 /// Client-visible rejection categories. Each member maps to exactly one status and fixed message in
