+++ SSO-Auth/Api/Shared/FlowResponses.cs
@@ -3,6 +3,7 @@ using System.Net.Mime;
 using System.Security.Cryptography;
 using Jellyfin.Plugin.SSO_Auth.Api;
 using Jellyfin.Plugin.SSO_Auth.Api.Linking;
+using Jellyfin.Plugin.SSO_Auth.Api.Session;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
