+++ SSO-Auth/Config/SsoOnlyLoginGuard.cs
@@ -1,5 +1,6 @@
 using System;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Session;
 
 namespace Jellyfin.Plugin.SSO_Auth.Config;
